refactoring

+++ src/main/java/com/example/service/storage/AbstractResourceService.java
@@ -1,13 +1,16 @@
 package com.example.service.storage;
 
 import com.example.dto.ResourceInfoDto;
-import com.example.exception.ResourceNotFoundException;
+import com.example.dto.enums.ResourceType;
+import com.example.exception.resource.ResourceException;
+import com.example.exception.resource.ResourceNotFoundException;
+import com.example.exception.resource.ResourceTypeException;
 import com.example.mapper.ResourceInfoMapper;
 import com.example.models.ResourceInfo;
 import com.example.models.StorageKey;
 import com.example.service.FileMetadataService;
 import com.example.service.ResourceService;
-import com.example.service.StorageService;
+import com.example.service.minio.StorageService;
 import io.minio.StatObjectResponse;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -48,10 +51,20 @@ public abstract class AbstractResourceService implements ResourceService {
 
     @Transactional
     private ResourceInfoDto upload(StorageKey storageKey, MultipartFile file) {
+        validateResource(storageKey, ResourceType.DIRECTORY);
+        if (fileMetadataService.isFilePresented(storageKey.getKey(), storageKey.getPath(), file.getName())) {
+            throw new ResourceException("Couldn't upload file, because file with path [%s] and name [%s] already exist".formatted(storageKey.getPath(), file.getName()));
+        }
         storageService.putObject(storageKey, file);
         StatObjectResponse statObjectResponse = storageService.getStatObject(storageKey);
         StorageKey statObjectStorageKey = StorageKey.parsePath(statObjectResponse.object());
         fileMetadataService.save(statObjectStorageKey, statObjectResponse.size());
         return resourceInfoMapper.toResourceInfoDto(statObjectStorageKey, statObjectResponse.size());
+    }
+
+    protected void validateResource(StorageKey storageKey, ResourceType expectedResourceType) {
+        if (storageKey.getResourceType() != expectedResourceType) {
+            throw new ResourceTypeException("Wrong storage key type, expected type is [%s]".formatted(expectedResourceType));
+        }
     }
 }
